requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudscraper==1.2.58
colorama==0.4.4
future==0.18.2
idna==3.3
names==0.3.0
pyparsing==3.0.7
PySocks==1.7.1
pyuseragents==1.0.5
requests==2.27.1
requests-toolbelt==0.9.1
soupsieve==2.3.1
tableprint==0.9.1
urllib3==1.26.7
uuid==1.30
wcwidth==0.2.5
win32-setctime==1.1.0
wincertstore==0.2
pympler==1.0.1
python-dotenv==0.20.0
